pyproject.toml
[project]
name = "agent-medha"
version = "0.1.0"
description = "A smart, brainy personal AI assistant using the Supervisor Pattern."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-google-genai>=2.0.0",
    "langsmith>=0.1.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "qdrant-client>=1.11.0",
    "fastapi>=0.110.0",
    "uvicorn>=0.27.0",
    "google-generativeai>=0.8.5",
    "langfuse>=3.10.1",
    "ragas>=0.3.9",
    "datasets>=4.4.1",
    "tweepy>=4.14.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/agent_medha"]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "ruff>=0.6.0",
    "mypy>=1.11.0",
    "ipykernel>=6.29.0",
]
